Hash user passkey in torrent info. Fixes #2.
Before this change, a user's passkey was stored in plaintext inside all
of his downloaded torrent files. If an attacker was able to steal any of
these files, he would then be able to download any torrent on behalf of
that user.

Now the torrent ID and the user's passkey are hashed before they are
stored in the torrent info. This hash can only be used to download this
particular torrent on behalf of that user.

This change breaks all existing torrent files. For security, all user
passkeys should be regenerated (securely) after this change has been
deployed.

Thanks to @vipera7 for taking an initial stab at this in #4 and #5.